'add CPU speed as metadata' (TrigCostMonitor-01-17-00)
	* Add extra meta info on CPU usage, feature req.
	* Tag TrigCostMonitor-01-17-00

2015-03-13 Tim Martin <Tim.Martin@cern.ch>
	* By default don't write out non-cost-events
	* Decided to duplicate xAOD::StreamTags to EventInfo::TriggerInfo for legacy support
	* (these two are combined later, and we need to make sure the cost stream is _removed_)
	* Tag TrigCostMonitor-01-15-00

2015-03-09 Tim Martin <Tim.Martin@cern.ch>
	* Reduce output of TrigCostRun yet more
	* Change to only fill xAOD::StreamTag
	* Tag TrigCostMonitor-01-14-00

+++ Trigger/TrigMonitoring/TrigCostMonitor/src/TrigNtConfTool.cxx
@@ -208,7 +208,7 @@ bool Trig::TrigNtConfTool::GetKeysChangedFromSv(TrigMonConfig &confg) {
    if ( m_configSvc->masterKey() != confg.getMasterKey() ) {
      ATH_MSG_DEBUG("Master key has changed from " << confg.getMasterKey() << " to " << m_configSvc->masterKey() << " - (re)fill " );
      return true;
-   } else if ( ctp_confg->prescaleSetId() != confg.getLV1PrescaleKey() ) {
+   } else if ( (int)ctp_confg->prescaleSetId() != (int)confg.getLV1PrescaleKey() ) {
      ATH_MSG_DEBUG("L1 key has changed from " << confg.getMasterKey() << " to " << ctp_confg->prescaleSetId() << " - (re)fill " );
      return true;
    } else  if ( m_configSvc->hltPrescaleKey() != confg.getHLTPrescaleKey() ) {
